requirements.txt
aiohttp-client-cache[sqlite]>=0.12.4
fake-useragent>=2.0.3
fastapi>=0.114.0
pydantic>=2.9.1
python-dotenv>=1.0.1
uvicorn>=0.30.6
yt-dlp
requests
pillow